requirements.txt
requests==2.25.1
schedule==0.6.0
websocket-client==0.57.0
PyQt5